+++ examples/GameOfLife/src/GameOfLife3.jl
@@ -50,10 +50,10 @@ function transition(c::Cell, id, sim)
 end
 
 const model = ModelTypes() |>    
-    add_agenttype!(Cell, :Vector; size = 40000) |> 
-#    add_agenttype!(Cell) |> 
-    add_edgetype!(Neighbor, :Stateless, :SingleAgentType; to_agenttype = Cell, size=40000) |>
-    add_edgetype!(ActiveNeighbor, :Stateless, :IgnoreFrom, :SingleAgentType; to_agenttype = Cell , size=40000) |>
+    register_agenttype!(Cell, :Vector; size = 40000) |> 
+#    register_agenttype!(Cell) |> 
+    register_edgetype!(Neighbor, :Stateless, :SingleAgentType; to_agenttype = Cell, size=40000) |>
+    register_edgetype!(ActiveNeighbor, :Stateless, :IgnoreFrom, :SingleAgentType; to_agenttype = Cell , size=40000) |>
     construct_model("GameOfLife")