+++ src/FactoryHelpers.jl
@@ -1,7 +1,1 @@
 
-writefield(T) = Symbol(T, "_write")
-
-readfield(T) = Symbol(T, "_read")
-
-nextidfield(T) = Symbol(T, "_nextid")
-